# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Order API",
        "contact": {
            "email": "arvind.gupta01@nagarro.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "L5 Swagger OpenApi Server"
        }
    ],
    "paths": {
        "/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get list of Orders",
                "description": "Returns list of orders",
                "operationId": "getOrdersList",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "[{'id':1,'distance':10,'status':'UNASSIGNED'},{'id':2,'distance':10,'status':'UNASSIGNED'}]"
                    },
                    "422": {
                        "description": "{'error':'The limit must be an integer.'}"
                    }
                }
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Place Order",
                "description": "Create a new Order",
                "operationId": "StoreOrder",
                "requestBody": {
                    "description": "Post object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "origin": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "destination": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "origin": [
                                        "40.6655101",
                                        "-73.89188969999998"
                                    ],
                                    "destination": [
                                        "40.6905615",
                                        "-73.9976592"
                                    ]
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "{'id':17,'distance':10434,'status':'UNASSIGNED'}"
                    },
                    "422": {
                        "description": "{'error':'Error message'}"
                    }
                }
            }
        },
        "/orders/{id}": {
            "patch": {
                "tags": [
                    "Orders"
                ],
                "summary": "Take Order",
                "description": "Take order",
                "operationId": "UpdateOrder",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "OrderId",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "status": "TAKEN"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "{\"status\":\"SUCCESS\"}"
                    },
                    "422": {
                        "description": "{'error':'Error message'}"
                    },
                    "404": {
                        "description": "{'error':'404 not found'}"
                    }
                }
            }
        }
    }
}